+++ src/libstore-test-support/include/nix/store/tests/libstore.hh
@@ -5,6 +5,7 @@
 #include <gmock/gmock.h>
 
 #include "nix/store/store-api.hh"
+#include "nix/store/store-open.hh"
 
 namespace nix {
