+++ src/libstore/include/nix/store/meson.build
@@ -66,16 +66,18 @@ headers = [config_pub_h] + files(
   'restricted-store.hh',
   's3-binary-cache-store.hh',
   's3.hh',
-  'ssh-store.hh',
   'serve-protocol-connection.hh',
   'serve-protocol-impl.hh',
   'serve-protocol.hh',
   'sqlite.hh',
+  'ssh-store.hh',
   'ssh.hh',
   'store-api.hh',
   'store-cast.hh',
   'store-dir-config.hh',
+  'store-open.hh',
   'store-reference.hh',
+  'store-registration.hh',
   'uds-remote-store.hh',
   'worker-protocol-connection.hh',
   'worker-protocol-impl.hh',
