+++ src/libstore/include/nix/store/store-api.hh
@@ -889,99 +889,6 @@ OutputPathMap resolveDerivedPath(Store &, const DerivedPath::Built &, Store * ev
 
 
 /**
- * @return a Store object to access the Nix store denoted by
- * ‘uri’ (slight misnomer...).
- */
-ref<Store> openStore(StoreReference && storeURI);
-
-
-/**
- * Opens the store at `uri`, where `uri` is in the format expected by `StoreReference::parse`
-
- */
-ref<Store> openStore(const std::string & uri = settings.storeUri.get(),
-    const Store::Config::Params & extraParams = Store::Config::Params());
-
-
-/**
- * @return the default substituter stores, defined by the
- * ‘substituters’ option and various legacy options.
- */
-std::list<ref<Store>> getDefaultSubstituters();
-
-struct StoreFactory
-{
-    /**
-     * Documentation for this type of store.
-     */
-    std::string doc;
-
-    /**
-     * URIs with these schemes should be handled by this factory
-     */
-    StringSet uriSchemes;
-
-    /**
-     * An experimental feature this type store is gated, if it is to be
-     * experimental.
-     */
-    std::optional<ExperimentalFeature> experimentalFeature;
-
-    /**
-     * The `authorityPath` parameter is `<authority>/<path>`, or really
-     * whatever comes after `<scheme>://` and before `?<query-params>`.
-     */
-    std::function<ref<StoreConfig>(
-        std::string_view scheme, std::string_view authorityPath, const Store::Config::Params & params)>
-        parseConfig;
-
-    /**
-     * Just for dumping the defaults. Kind of awkward this exists,
-     * because it means we cannot require fields to be manually
-     * specified so easily.
-     */
-    std::function<ref<StoreConfig>()> getConfig;
-};
-
-struct Implementations
-{
-    using Map = std::map<std::string, StoreFactory>;
-
-    static Map & registered();
-
-    template<typename TConfig>
-    static void add()
-    {
-        StoreFactory factory{
-            .doc = TConfig::doc(),
-            .uriSchemes = TConfig::uriSchemes(),
-            .experimentalFeature = TConfig::experimentalFeature(),
-            .parseConfig =
-                ([](auto scheme, auto uri, auto & params)
-                 -> ref<StoreConfig>
-                 { return make_ref<TConfig>(scheme, uri, params); }),
-            .getConfig =
-                ([]() -> ref<StoreConfig>
-                 { return make_ref<TConfig>(Store::Config::Params{}); }),
-        };
-        auto [it, didInsert] = registered().insert({TConfig::name(), std::move(factory)});
-        if (!didInsert) {
-            throw Error("Already registred store with name '%s'", it->first);
-        }
-    }
-};
-
-template<typename TConfig>
-struct RegisterStoreImplementation
-{
-    RegisterStoreImplementation()
-    {
-        Implementations::add<TConfig>();
-    }
-};
-
-
-/**
  * Display a set of paths in human-readable form (i.e., between quotes
  * and separated by commas).
  */
