+++ src/libstore/uds-remote-store.cc
@@ -1,6 +1,7 @@
 #include "nix/store/uds-remote-store.hh"
 #include "nix/util/unix-domain-socket.hh"
 #include "nix/store/worker-protocol.hh"
+#include "nix/store/store-registration.hh"
 
 #include <sys/types.h>
 #include <sys/stat.h>
