+++ src/nix-build/nix-build.cc
@@ -12,7 +12,7 @@
 #include "nix/util/current-process.hh"
 #include "nix/store/parsed-derivations.hh"
 #include "nix/store/derivation-options.hh"
-#include "nix/store/store-api.hh"
+#include "nix/store/store-open.hh"
 #include "nix/store/local-fs-store.hh"
 #include "nix/store/globals.hh"
 #include "nix/store/realisation.hh"
