+++ src/nix-channel/nix-channel.cc
@@ -2,7 +2,7 @@
 #include "nix/main/shared.hh"
 #include "nix/store/globals.hh"
 #include "nix/store/filetransfer.hh"
-#include "nix/store/store-api.hh"
+#include "nix/store/store-open.hh"
 #include "nix/cmd/legacy.hh"
 #include "nix/expr/eval-settings.hh" // for defexpr
 #include "nix/util/users.hh"
