+++ src/nix/make-content-addressed.cc
@@ -1,5 +1,5 @@
 #include "nix/cmd/command.hh"
-#include "nix/store/store-api.hh"
+#include "nix/store/store-open.hh"
 #include "nix/store/make-content-addressed.hh"
 #include "nix/main/common-args.hh"
