+++ src/nix/verify.cc
@@ -1,6 +1,6 @@
 #include "nix/cmd/command.hh"
 #include "nix/main/shared.hh"
-#include "nix/store/store-api.hh"
+#include "nix/store/store-open.hh"
 #include "nix/util/thread-pool.hh"
 #include "nix/util/signals.hh"
 #include "nix/store/keys.hh"
