+++ tests/functional/test-libstoreconsumer/main.cc
@@ -1,5 +1,5 @@
 #include "nix/store/globals.hh"
-#include "nix/store/store-api.hh"
+#include "nix/store/store-open.hh"
 #include "nix/store/build-result.hh"
 #include <iostream>
